Fix Identation, Replace NGI reference with superGroup parameter
