+++ status-computation/java/src/main/java/ar/GroupEndpointIntegrate.java
@@ -20,28 +20,29 @@ import org.eclipse.jdt.core.dom.ThisExpression;
 import sync.AvailabilityProfiles;
 import sync.GroupsOfGroups;
 
-public class SiteIntegrate extends EvalFunc<Tuple> {
+public class GroupEndpointIntegrate extends EvalFunc<Tuple> {
 
 	public AvailabilityProfiles apMgr;
 	public GroupsOfGroups ggMgr;
 	public OpsManager opsMgr;
-	
-	
+
 	private String fnAps;
 	public String fnGroups;
-	
+
 	public String fnOps;
-	
-	
+
 	private String fsUsed;
 	public DIntegrator arMgr;
-	
+
 	private boolean initialized = false;
-	
-	private TupleFactory tupFactory; 
-    private BagFactory bagFactory;
-	
-	public SiteIntegrate(String fnOps, String fnAps, String fnGroups, String fsUsed) {
+
+	private TupleFactory tupFactory;
+	private BagFactory bagFactory;
+
+	private String superGroup;
+
+	public GroupEndpointIntegrate(String fnOps, String fnAps, String fnGroups,
+			String fsUsed, String superGroup) {
 		this.fnAps = fnAps;
 		this.fsUsed = fsUsed;
 		this.fnOps = fnOps;
@@ -50,73 +51,72 @@ public class SiteIntegrate extends EvalFunc<Tuple> {
 		this.ggMgr = new GroupsOfGroups();
 		this.opsMgr = new OpsManager();
 		this.arMgr = new DIntegrator();
-	
+		this.superGroup = superGroup;
+
 		this.tupFactory = TupleFactory.getInstance();
 		this.bagFactory = BagFactory.getInstance();
 	}
-	
-	public void init() throws IOException
-	{
-		if (this.fsUsed.equalsIgnoreCase("cache")){
+
+	public void init() throws IOException {
+		if (this.fsUsed.equalsIgnoreCase("cache")) {
 			this.apMgr.loadProfileJson(new File("./aps"));
 			this.ggMgr.loadAvro(new File("./groups"));
 			this.opsMgr.openFile(new File("./ops"));
 		}
-		
-		this.initialized=true;
-		
+
+		this.initialized = true;
+
 	}
-	
-	public List<String> getCacheFiles() { 
-        List<String> list = new ArrayList<String>(); 
-        list.add(this.fnAps.concat("#aps"));
-        list.add(this.fnGroups.concat("#groups"));
-        list.add(this.fnOps.concat("#ops"));
-        return list; 
-	} 
-	
+
+	public List<String> getCacheFiles() {
+		List<String> list = new ArrayList<String>();
+		list.add(this.fnAps.concat("#aps"));
+		list.add(this.fnGroups.concat("#groups"));
+		list.add(this.fnOps.concat("#ops"));
+		return list;
+	}
+
 	@Override
 	public Tuple exec(Tuple input) throws IOException {
-		
-		// Check if cache files have been opened 
-		if (this.initialized==false)
-        {
-        	this.init(); // If not open them 
-        }
-		
-		if (input == null || input.size() == 0) return null;
-		
-		String sitename = (String) input.get(0);
-		
-		//Get the Timeline
-		DTimeline siteTl = new DTimeline();
-		
-		DefaultDataBag bag =  (DefaultDataBag) input.get(1);
-		Iterator<Tuple> itBag = bag.iterator();
-		int j=0;
-		while (itBag.hasNext()){
-	    	Tuple curItem = itBag.next();
-	    	siteTl.samples[j] = Integer.parseInt(curItem.get(0).toString());
-	    	j++;
+
+		// Check if cache files have been opened
+		if (this.initialized == false) {
+			this.init(); // If not open them
 		}
-		
+
+		if (input == null || input.size() == 0)
+			return null;
+
+		String groupname = (String) input.get(0);
+
+		// Get the Timeline
+		DTimeline siteTl = new DTimeline();
+
+		DefaultDataBag bag = (DefaultDataBag) input.get(1);
+		Iterator<Tuple> itBag = bag.iterator();
+		int j = 0;
+		while (itBag.hasNext()) {
+			Tuple curItem = itBag.next();
+			siteTl.samples[j] = Integer.parseInt(curItem.get(0).toString());
+			j++;
+		}
+
 		this.arMgr.calculateAR(siteTl.samples, this.opsMgr);
-		
+
 		String aprofile = this.apMgr.getAvProfiles().get(0);
-		
+
 		Tuple output = tupFactory.newTuple();
 		output.append(this.apMgr.getProfileMetricProfile(aprofile));
 		output.append(aprofile);
-		output.append(sitename);
-		output.append(this.ggMgr.getGroup("NGI", sitename)); // NGI
+		output.append(groupname);
+		output.append(this.ggMgr.getGroup(superGroup, groupname));
 		output.append(this.arMgr.availability);
 		output.append(this.arMgr.reliability);
 		output.append(this.arMgr.up_f);
 		output.append(this.arMgr.unknown_f);
 		output.append(this.arMgr.down_f);
-		
+
 		return output;
 	}
-		
-		
+
 }
